+++ OftenOn/DSCResources/ooRegistry/ooRegistry.schema.psm1
@@ -42,13 +42,11 @@ Configuration ooRegistry {
         ValueType = 'DWord'
     }
 
-    # When the network location awareness service starts up before the domain starts (especially on Windows 2016),
-    # the network adapters get changed to public instead of domain, and then everything fails
-    Registry 'EnableNetworkLocationAwarenessServiceDelayedStart' {
+    Registry 'DisableInactivityTimeout' {
         Ensure    = 'Present'
-        Key       = 'HKEY_LOCAL_MACHINE\SYSTEM\CurrentControlSet\services\nlasvc'
-        ValueName = 'DelayedAutoStart'
-        ValueData = '1'
+        Key       = 'HKEY_LOCAL_MACHINE\Software\Microsoft\Windows\CurrentVersion\Policies\System'
+        ValueName = 'InactivityTimeoutSecs'
+        ValueData = '0'
         ValueType = 'DWord'
     }
 }