+++ OftenOn/OftenOn.psm1
@@ -1,7 +1,7 @@
 Set-StrictMode -Version Latest
 $ErrorActionPreference = "Stop"
 
-Get-ChildItem $PSScriptRoot *.ps1 -Exclude *.Tests.ps1 -Recurse | ForEach-Object {
+Get-ChildItem $PSScriptRoot *.ps1 -Exclude NlaSvcFix.ps1, TriggerDsc.ps1, *.Tests.ps1 -Recurse | ForEach-Object {
     Write-Verbose "Loading $($PSItem.FullName)"
     . $PSItem.FullName
 }